Because only detail list view need the background, so move the logic into the child class.
